+++ include/envoy/network/listen_socket.h
@@ -6,6 +6,8 @@
 #include "envoy/common/pure.h"
 #include "envoy/network/address.h"
 
+#include "absl/strings/string_view.h"
+
 namespace Envoy {
 namespace Network {
 
@@ -121,6 +123,26 @@ public:
    *         address the socket was initially accepted at.
    */
   virtual bool localAddressRestored() const PURE;
+
+  /**
+   * Set detected transport protocol (e.g. RAW_BUFFER, TLS).
+   */
+  virtual void setDetectedTransportProtocol(absl::string_view protocol) PURE;
+
+  /**
+   * @return detected transport protocol (e.g. RAW_BUFFER, TLS), if any.
+   */
+  virtual absl::string_view detectedTransportProtocol() const PURE;
+
+  /**
+   * Set requested server name (e.g. SNI in TLS).
+   */
+  virtual void setRequestedServerName(absl::string_view server_name) PURE;
+
+  /**
+   * @return requested server name (e.g. SNI in TLS), if any.
+   */
+  virtual absl::string_view requestedServerName() const PURE;
 };
 
 typedef std::unique_ptr<ConnectionSocket> ConnectionSocketPtr;