+++ test/common/router/config_impl_test.cc
@@ -4214,6 +4214,9 @@ class PerFilterConfigsTest : public testing::Test {
 public:
   PerFilterConfigsTest() : factory_(), registered_factory_(factory_) {}
 
+  struct DerivedFilterConfig : public RouteSpecificFilterConfig {
+    ProtobufWkt::Timestamp config_;
+  };
   class TestFilterConfig : public Extensions::HttpFilters::Common::EmptyHttpFilterConfig {
   public:
     Server::Configuration::HttpFilterFactoryCb
@@ -4222,6 +4225,12 @@ public:
     }
     ProtobufTypes::MessagePtr createEmptyRouteConfigProto() override {
       return ProtobufTypes::MessagePtr{new ProtobufWkt::Timestamp()};
+    }
+    Router::RouteSpecificFilterConfigConstSharedPtr
+    createRouteSpecificFilterConfig(const Protobuf::Message& message) override {
+      auto obj = std::make_shared<DerivedFilterConfig>();
+      obj->config_.MergeFrom(message);
+      return obj;
     }
     std::string name() override { return "test.filter"; }
   };
@@ -4234,19 +4243,18 @@ public:
     const auto* route_entry = route->routeEntry();
     const auto& vhost = route_entry->virtualHost();
 
-    check(route_entry->perFilterConfig(factory_.name()), expected_entry, "route entry");
-    check(route->perFilterConfig(factory_.name()), expected_route, "route");
-    check(vhost.perFilterConfig(factory_.name()), expected_vhost, "virtual host");
+    check(route_entry->perFilterConfigTyped<DerivedFilterConfig>(factory_.name()), expected_entry,
+          "route entry");
+    check(route->perFilterConfigTyped<DerivedFilterConfig>(factory_.name()), expected_route,
+          "route");
+    check(vhost.perFilterConfigTyped<DerivedFilterConfig>(factory_.name()), expected_vhost,
+          "virtual host");
   }
 
-  void check(const Protobuf::Message* cfg, uint32_t expected_seconds, std::string source) {
+  void check(const DerivedFilterConfig* cfg, uint32_t expected_seconds, std::string source) {
     EXPECT_NE(nullptr, cfg) << "config should not be null for source: " << source;
-    EXPECT_NO_THROW({
-      const auto ts = dynamic_cast<const ProtobufWkt::Timestamp*>(cfg);
-      EXPECT_EQ(expected_seconds, ts->seconds())
-          << "config value does not match expected for source: " << source;
-    }) << "config should properly dynamic_cast to the appropriate type for source: "
-       << source;
+    EXPECT_EQ(expected_seconds, cfg->config_.seconds())
+        << "config value does not match expected for source: " << source;
   }
 
   TestFilterConfig factory_;