+++ test/server/listener_socket_option_impl_test.cc
@@ -1,22 +1,27 @@
 #include "server/listener_socket_option_impl.h"
 
-#include "test/common/network/socket_option_test_harness.h"
+#include "test/common/network/socket_option_test.h"
+#include "test/test_common/environment.h"
 
 namespace Envoy {
 namespace Server {
 
-class ListenerSocketOptionImplTest : public Network::SocketOptionTestHarness {
+class ListenerSocketOptionImplTest : public Network::SocketOptionTest {
 public:
   void testSetSocketOptionSuccess(ListenerSocketOptionImpl& socket_option, int socket_level,
                                   Network::SocketOptionName option_name, int option_val,
                                   const std::set<Network::Socket::SocketState>& when) {
-    Network::SocketOptionTestHarness::testSetSocketOptionSuccess(socket_option, socket_level,
-                                                                 option_name, option_val, when);
+    Network::SocketOptionTest::testSetSocketOptionSuccess(socket_option, socket_level, option_name,
+                                                          option_val, when);
   }
 };
 
+INSTANTIATE_TEST_CASE_P(IpVersions, ListenerSocketOptionImplTest,
+                        testing::ValuesIn(TestEnvironment::getIpVersionsForTest()),
+                        TestUtility::ipTestParamsToString);
+
 // We fail to set the tcp-fastopen option when the underlying setsockopt syscall fails.
-TEST_F(ListenerSocketOptionImplTest, SetOptionTcpFastopenFailure) {
+TEST_P(ListenerSocketOptionImplTest, SetOptionTcpFastopenFailure) {
   if (ENVOY_SOCKET_TCP_FASTOPEN.has_value()) {
     ListenerSocketOptionImpl socket_option{{}, {}, 1};
     EXPECT_CALL(os_sys_calls_, setsockopt_(_, IPPROTO_TCP, ENVOY_SOCKET_TCP_FASTOPEN.value(), _, _))
@@ -26,14 +31,14 @@ TEST_F(ListenerSocketOptionImplTest, SetOptionTcpFastopenFailure) {
 }
 
 // The happy path for setOption(); TCP_FASTOPEN is set to true.
-TEST_F(ListenerSocketOptionImplTest, SetOptionTcpFastopenSuccessTrue) {
+TEST_P(ListenerSocketOptionImplTest, SetOptionTcpFastopenSuccessTrue) {
   ListenerSocketOptionImpl socket_option{{}, {}, 42};
   testSetSocketOptionSuccess(socket_option, IPPROTO_TCP, ENVOY_SOCKET_TCP_FASTOPEN, 42,
                              {Network::Socket::SocketState::Listening});
 }
 
 // The happy path for setOption(); TCP_FASTOPEN is set to false.
-TEST_F(ListenerSocketOptionImplTest, SetOptionTcpFastopenSuccessFalse) {
+TEST_P(ListenerSocketOptionImplTest, SetOptionTcpFastopenSuccessFalse) {
   ListenerSocketOptionImpl socket_option{{}, {}, 0};
   testSetSocketOptionSuccess(socket_option, IPPROTO_TCP, ENVOY_SOCKET_TCP_FASTOPEN, 0,
                              {Network::Socket::SocketState::Listening});